pyproject.toml
# Copyright DB Netz AG and contributors
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = ["setuptools>=64", "setuptools_scm[toml]>=3.4", "wheel"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]

name = "capella2polarion"
description = "Synchronise Capella models with Polarion projects"
readme = "README.md"
requires-python = ">=3.9, <3.12"
license = { text = "Apache-2.0" }
authors = [
  { name = "DB Netz AG" },
]
keywords = []
classifiers = [
  "Development Status :: 1 - Planning",
  "License :: OSI Approved :: Apache Software License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]
dependencies = [
  "capellambse",
  "click",
  "PyYAML",
  "polarion-rest-api-client",
  "requests",
]

[project.urls]
Homepage = "https://github.com/DSD-DBS/capella2polarion"
Documentation = "https://dsd-dbs.github.io/capella2polarion"

[project.optional-dependencies]
dev = [
  "python-dotenv"
]

docs = [
  "furo",
  "sphinx",
  "sphinx-copybutton",
  "tomli; python_version<'3.11'",
]

test = [
  "pytest",
  "pytest-cov",
  "pytest-httpx",
]

[tool.black]
line-length = 79
target-version = ["py310"]

[tool.docformatter]
wrap-descriptions = 72
wrap-summaries = 79

[tool.isort]
profile = 'black'
line_length = 79

[tool.mypy]
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_redundant_casts = true
warn_unreachable = true
python_version = "3.10"

[[tool.mypy.overrides]]
module = ["tests.*"]
allow_incomplete_defs = true
allow_untyped_defs = true

[[tool.mypy.overrides]]
# Untyped third party libraries
module = [
  # ...
]
ignore_missing_imports = true

[tool.pydocstyle]
convention = "numpy"
add-select = [
  "D212",  # Multi-line docstring summary should start at the first line
  "D402",  # First line should not be the function’s “signature”
  "D417",  # Missing argument descriptions in the docstring
]
add-ignore = [
  "D201",  # No blank lines allowed before function docstring  # auto-formatting
  "D202",  # No blank lines allowed after function docstring  # auto-formatting
  "D203",  # 1 blank line required before class docstring  # auto-formatting
  "D204",  # 1 blank line required after class docstring  # auto-formatting
  "D211",  # No blank lines allowed before class docstring  # auto-formatting
  "D213",  # Multi-line docstring summary should start at the second line
]

[tool.pylint.master]
extension-pkg-allow-list = [
  "lxml.etree",
]
max-line-length = 79

[tool.pylint.messages_control]
disable = [
  "broad-except",
  "global-statement",
  "import-outside-toplevel",
  "invalid-name",
  "missing-class-docstring",
  "missing-function-docstring",
  "missing-module-docstring",
  "no-else-break",
  "no-else-continue",
  "no-else-raise",
  "no-else-return",
  "protected-access",
  "redefined-builtin",
  "too-few-public-methods",
  "too-many-ancestors",
  "too-many-arguments",
  "too-many-boolean-expressions",
  "too-many-branches",
  "too-many-instance-attributes",
  "too-many-lines",
  "too-many-locals",
  "too-many-public-methods",
  "too-many-return-statements",
  "too-many-statements",

  # Auto-formatting
  "bad-indentation",
  "inconsistent-quotes",
  "missing-final-newline",
  "mixed-line-endings",
  "multiple-imports",
  "multiple-statements",
  "trailing-newlines",
  "trailing-whitespace",
  "unexpected-line-ending-format",
  "ungrouped-imports",
  "wrong-import-order",
  "wrong-import-position",

  # Handled by mypy
  "arguments-differ",
  "assignment-from-no-return",
  "import-error",
  "missing-kwoa",
  "no-member",
  "no-value-for-parameter",
  "redundant-keyword-arg",
  "signature-differs",
  "syntax-error",
  "too-many-function-args",
  "unbalanced-tuple-unpacking",
  "undefined-variable",
  "unexpected-keyword-arg",
]
enable = [
  "c-extension-no-member",
  "deprecated-pragma",
  "use-symbolic-message-instead",
  "useless-suppression",
]

[tool.pytest.ini_options]
addopts = """
  --import-mode=importlib
  --strict-config
  --strict-markers
"""
testpaths = ["tests"]
xfail_strict = true

[tool.setuptools]
platforms = ["any"]
zip-safe = false

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.setuptools.packages.find]
exclude = ["LICENSES"]

[tool.setuptools_scm]
# This section must exist for setuptools_scm to work
